+++ R/mirai.R
@@ -78,6 +78,9 @@
 #'     \code{\link{mirai}} is an alias for \code{\link{eval_mirai}}.
 #'
 #' @examples
+#' if (interactive()) {
+#' # Only run examples in interactive R sessions
+#'
 #' m <- mirai(x + y + 1, x = 2, y = 3)
 #' m
 #' m$data
@@ -96,6 +99,8 @@
 #'   Sys.sleep(0.1)
 #' }
 #' str(m$data)
+#'
+#' }
 #'
 #' @export
 #'